internal: move all struct* packages into libs/structs/
