+++ libs/structs/structdiff/bench.sh
