+++ libs/structs/structdiff/bench_test.go
