+++ libs/structs/structdiff/jobsettings_test.go
