+++ libs/structs/structpath/path_test.go
