+++ libs/structs/structwalk/util_test.go
