+++ libs/structs/structwalk/walk.go
@@ -6,7 +6,7 @@ import (
 	"slices"
 	"sort"
 
-	"github.com/databricks/cli/libs/structdiff/structpath"
+	"github.com/databricks/cli/libs/structs/structpath"
 )
 
 // VisitFunc is invoked for every scalar (int, uint, float, string, bool) field encountered while walking v.
