+++ libs/structs/structwalk/walk_test.go
@@ -3,7 +3,7 @@ package structwalk
 import (
 	"testing"
 
-	"github.com/databricks/cli/libs/structdiff/structpath"
+	"github.com/databricks/cli/libs/structs/structpath"
 	"github.com/databricks/databricks-sdk-go/service/jobs"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
